+++ terminal/backends/__init__.py
@@ -1,12 +1,15 @@
 from .pipe_backend import PipeProcessSession
 from .pty_backend import PtyProcessSession, UnsupportedPtySession
+from .sshpass_backend import SshpassPromptSession, can_handle_sshpass
 from .tmux_backend import TmuxSession
 from .winpty_backend import WinPtySession
 
 __all__ = [
     "PipeProcessSession",
     "PtyProcessSession",
+    "SshpassPromptSession",
     "TmuxSession",
     "UnsupportedPtySession",
     "WinPtySession",
+    "can_handle_sshpass",
 ]